fix: debugged registration function to work and hashed with salt the passwords using Spring Security

+++ src/main/java/com/example/user/domain/repository/UserRepository.java
@@ -8,7 +8,7 @@ public interface UserRepository {
 
     User findUserByEmail(String email);
 
-    long createUser(User user);
+    Integer createUser(User user);
 
     void removeUser(String email);
